rename macro REG_SP() to REG_GET_SP()
This change is applied to all existing architectures as the function in the generic code is renamed. The reason is for riscv64, REG_SP macro is already defined in glibc to refer to the register itself (`#define REG_SP 2` in `glibc/sysdeps/unix/sysv/linux/riscv/sys/ucontext.h`). Thus we renamed REG_SP to avoid duplicated definitions.

Signed-off-by: Yixue Zhao <felicitia2010@gmail.com>
Co-Authored-By: Rushikesh Butley <rushikeshbutley@gmail.com>